오류발견 in productdao.java

+++ src/main/java/com/onsil/onsil/admin/dto/PopularCountDto.java
@@ -11,6 +11,6 @@ public class PopularCountDto {
 
     private Integer productId;
     private String flowerName;
-    private long subscribeCount;
+    private Integer subscribeCount;
 
 }
